+++ changelog.md
@@ -2,7 +2,9 @@
 
 * Added kanji references for Conning's 'The Kodansha Kanji Learner's Course'
   thanks to [@Kala-J](https://github.com/Kala-J).
-* Make the kanji view show the traditional radical (not the Nelson radical)
+* Make the kanji view show the traditional radical (not the Nelson radical).
+  The Nelson radical will be shown in the references section in cases where it
+  differs.
 
 ## 0.0.28 (2018-12-31